requirements.txt
win32com.client
pyinstaller